requirements.txt
flask
pinecone
python-dotenv
pandas
numpy
scikit-learn
sentence-transformers
pytesseract
Pillow
opencv-python
requests
beautifulsoup4
selenium
webdriver-manager
lxml
tensorflow


